Consolidate annotation test modules
