--- tests/dataclass_classvar.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-from typing import ClassVar
-
-@dataclass
-class Example:
-    x: int
-    y: ClassVar[int] = 0
