--- tests/dataclass_classvar.pyi
@@ -1,7 +1,0 @@
-from typing import ClassVar
-from dataclasses import dataclass
-
-@dataclass
-class Example:
-    x: int
-    y: ClassVar[int]
