--- tests/elliptic_tuple.py
@@ -1,4 +1,0 @@
-from typing import Tuple
-
-def use_tuple(tp: tuple[int, ...]) -> tuple[int, ...]:
-    return tp
