--- tests/elliptic_tuple.pyi
@@ -1,1 +1,0 @@
-def use_tuple(tp: tuple[int, ...]) -> tuple[int, ...]: ...
