--- tests/enums.pyi
@@ -1,10 +1,0 @@
-from enum import Enum, IntEnum
-
-class Color(Enum):
-    RED = 1
-    GREEN = 2
-
-class Priority(IntEnum):
-    LOW = 1
-    MEDIUM = 2
-    HIGH = 3
