--- tests/namedtuple_sample.py
@@ -1,5 +1,0 @@
-from typing import NamedTuple
-
-class Point(NamedTuple):
-    x: int
-    y: int
