--- tests/nested_dataclass.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Outer:
-    x: int
-    @dataclass
-    class Inner:
-        y: int
