--- tests/new_generic.py
@@ -1,8 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar('T')
-
-class NewGeneric[T]:
-    value: T
-    def get(self) -> T:
-        return self.value
