--- tests/new_generic.pyi
@@ -1,7 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar('T')
-
-class NewGeneric[T]:
-    value: T
-    def get(self) -> T: ...
