--- tests/old_generic.py
@@ -1,8 +1,0 @@
-from typing import TypeVar, Generic
-
-T = TypeVar('T')
-
-class OldGeneric(Generic[T]):
-    value: T
-    def get(self) -> T:
-        return self.value
