--- tests/old_generic.pyi
@@ -1,7 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar('T')
-
-class OldGeneric[T]:
-    value: T
-    def get(self) -> T: ...
