--- tests/self_annotation.py
@@ -1,5 +1,0 @@
-from typing import Self
-
-class Example:
-    def clone(self: Self) -> Self:
-        return self
