--- tests/self_annotation.pyi
@@ -1,4 +1,0 @@
-from typing import Self
-
-class Example:
-    def clone(self: Self) -> Self: ...
