+++ tests/test_mypy.py
@@ -6,7 +6,7 @@ from pathlib import Path
 def test_stub_files_pass_mypy():
     pyi_dir = Path(__file__).parent
     pyi_paths = sorted(pyi_dir.glob("*.pyi"))
-    skip = {"variadic.pyi", "new_generic.pyi", "old_generic.pyi"}
+    skip = {"annotations.pyi", "typechecking.pyi"}
     pyi_paths = [p for p in pyi_paths if p.name not in skip]
     for path in pyi_paths:
         result = subprocess.run(
