+++ tests/test_pyi_extract.py
@@ -9,128 +9,12 @@ from macrotype.stubgen import load_module_from_path
 
 
 def test_stub_generation_matches_expected():
-    src = Path(__file__).with_name("all_annotations.py")
+    src = Path(__file__).with_name("annotations.py")
     loaded = load_module_from_path(src)
     module = PyiModule.from_module(loaded)
     generated = module.render()
 
-    expected_path = Path(__file__).with_name("all_annotations.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_dataclass_decorators():
-    src = Path(__file__).with_name("dataclass_sample.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("dataclass_sample.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_nested_dataclass():
-    src = Path(__file__).with_name("nested_dataclass.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("nested_dataclass.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_variadic_typevar_tuple():
-    src = Path(__file__).with_name("variadic.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("variadic.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_new_style_generic_class():
-    src = Path(__file__).with_name("new_generic.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("new_generic.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_old_style_generic_class():
-    src = Path(__file__).with_name("old_generic.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("old_generic.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_tuple_ellipsis():
-    src = Path(__file__).with_name("elliptic_tuple.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("elliptic_tuple.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-def test_self_annotation():
-    src = Path(__file__).with_name("self_annotation.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("self_annotation.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-def test_enum_values():
-    src = Path(__file__).with_name("enums.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected = Path(__file__).with_name("enums.pyi").read_text().splitlines()
-    assert generated == expected
-
-
-def test_dataclass_classvar():
-    src = Path(__file__).with_name("dataclass_classvar.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("dataclass_classvar.pyi")
-    expected = expected_path.read_text().splitlines()
-
-    assert generated == expected
-
-
-def test_namedtuple_class():
-    src = Path(__file__).with_name("namedtuple_sample.py")
-    loaded = load_module_from_path(src)
-    module = PyiModule.from_module(loaded)
-    generated = module.render()
-
-    expected_path = Path(__file__).with_name("namedtuple_sample.pyi")
+    expected_path = Path(__file__).with_name("annotations.pyi")
     expected = expected_path.read_text().splitlines()
 
     assert generated == expected
