+++ tests/typechecking.pyi
@@ -1,3 +1,3 @@
-from tests.all_annotations import AllAnnotations
+from tests.annotations import Basic
 
-def takes(x: AllAnnotations) -> None: ...
+def takes(x: Basic) -> None: ...
